fix(security): remove renderer trust dependency from plugin security policy
Move plugin security policy to be exclusively main-process owned:
- Remove untrusted manifest registration path (registerManifest)
- Remove untrustedManifests map and getManifest fallback
- Move validateManifest to src/shared/ (was cross-layer import)
- Fix race condition: await async refreshManifest in process-handlers
- Remove REGISTER_MANIFEST IPC channel (renderer no longer sends manifests)
- Renderer uses refreshManifestFromDisk for community plugin registration

Closes #575
