+++ src/main/services/plugin-manifest-registry.test.ts
@@ -26,7 +26,6 @@ import { appLog } from './log-service';
 import {
   initializeTrustedManifests,
   refreshManifest,
-  registerManifest,
   registerTrustedManifest,
   getManifest,
   getAllowedCommands,
@@ -90,61 +89,30 @@ describe('plugin-manifest-registry', () => {
     });
   });
 
-  // ── Untrusted registration (renderer IPC) ──────────────────────────
+  // ── No untrusted registration path ─────────────────────────────────
 
-  describe('registerManifest (untrusted/IPC)', () => {
-    it('strips allowedCommands from renderer-sourced manifest', () => {
-      registerManifest('test-plugin', makeManifest({ allowedCommands: ['sh', 'bash', 'curl'] }));
-      expect(getAllowedCommands('test-plugin')).toEqual([]);
+  describe('getManifest returns only trusted manifests', () => {
+    it('returns undefined when no trusted manifest exists', () => {
+      expect(getManifest('unknown-plugin')).toBeUndefined();
     });
 
-    it('preserves non-sensitive fields', () => {
-      registerManifest('test-plugin', makeManifest({
-        name: 'Malicious Plugin',
-        version: '2.0.0',
-        allowedCommands: ['sh'],
-      }));
-      const manifest = getManifest('test-plugin');
-      expect(manifest?.name).toBe('Malicious Plugin');
-      expect(manifest?.version).toBe('2.0.0');
-      expect(manifest?.allowedCommands).toBeUndefined();
-    });
-
-    it('cannot overwrite trusted allowedCommands via untrusted re-registration', () => {
-      registerTrustedManifest('test-plugin', makeManifest({ allowedCommands: ['git'] }));
-      registerManifest('test-plugin', makeManifest({ allowedCommands: ['sh', 'bash', 'rm'] }));
-      expect(getAllowedCommands('test-plugin')).toEqual(['git']);
-    });
-
-    it('cannot register a fake plugin to gain allowedCommands', () => {
-      registerManifest('evil-plugin', makeManifest({
-        id: 'evil-plugin',
-        allowedCommands: ['sh', 'bash'],
-      }));
-      expect(getAllowedCommands('evil-plugin')).toEqual([]);
+    it('returns the trusted manifest', () => {
+      registerTrustedManifest('test-plugin', makeManifest({ name: 'Trusted' }));
+      expect(getManifest('test-plugin')?.name).toBe('Trusted');
     });
   });
 
-  // ── Self-escalation attack scenarios ─────────────────────────────────
+  // ── Security: no renderer path to inject policy ─────────────────────
 
-  describe('self-escalation prevention', () => {
-    it('blocks self-escalation: register new manifest then exec', () => {
-      registerManifest('malicious', makeManifest({
-        id: 'malicious',
-        permissions: ['process'],
-        allowedCommands: ['sh', 'bash', 'curl', 'rm'],
-      }));
+  describe('security invariants', () => {
+    it('getAllowedCommands returns empty for unregistered plugins (deny by default)', () => {
       expect(getAllowedCommands('malicious')).toEqual([]);
     });
 
-    it('blocks self-escalation: overwrite existing manifest', () => {
+    it('only trusted manifests can set allowedCommands', () => {
       registerTrustedManifest('my-plugin', makeManifest({
         id: 'my-plugin',
         allowedCommands: ['git'],
-      }));
-      registerManifest('my-plugin', makeManifest({
-        id: 'my-plugin',
-        allowedCommands: ['git', 'sh', 'bash', 'rm'],
       }));
       expect(getAllowedCommands('my-plugin')).toEqual(['git']);
     });
@@ -153,9 +121,8 @@ describe('plugin-manifest-registry', () => {
   // ── Utility functions ──────────────────────────────────────────────
 
   describe('unregisterManifest', () => {
-    it('removes both trusted and untrusted manifests', () => {
+    it('removes trusted manifest', () => {
       registerTrustedManifest('test-plugin', makeManifest());
-      registerManifest('test-plugin', makeManifest());
       expect(unregisterManifest('test-plugin')).toBe(true);
       expect(getManifest('test-plugin')).toBeUndefined();
     });
@@ -168,23 +135,10 @@ describe('plugin-manifest-registry', () => {
   describe('clear', () => {
     it('removes all manifests', () => {
       registerTrustedManifest('a', makeManifest({ id: 'a' }));
-      registerManifest('b', makeManifest({ id: 'b' }));
+      registerTrustedManifest('b', makeManifest({ id: 'b' }));
       clear();
       expect(getManifest('a')).toBeUndefined();
       expect(getManifest('b')).toBeUndefined();
-    });
-  });
-
-  describe('getManifest preference', () => {
-    it('prefers trusted manifest over untrusted', () => {
-      registerTrustedManifest('test-plugin', makeManifest({ name: 'Trusted' }));
-      registerManifest('test-plugin', makeManifest({ name: 'Untrusted' }));
-      expect(getManifest('test-plugin')?.name).toBe('Trusted');
-    });
-
-    it('falls back to untrusted manifest if no trusted exists', () => {
-      registerManifest('test-plugin', makeManifest({ name: 'Untrusted' }));
-      expect(getManifest('test-plugin')?.name).toBe('Untrusted');
     });
   });
 
@@ -242,7 +196,7 @@ describe('plugin-manifest-registry', () => {
     expect(getManifest('community-plugin')).toBeUndefined();
   });
 
-  it('refreshes a community manifest from disk instead of keeping stale renderer state', async () => {
+  it('refreshes a community manifest from disk instead of keeping stale state', async () => {
     vi.mocked(fs.readFileSync).mockReturnValue('true');
     vi.mocked(pluginDiscovery.discoverCommunityPlugins).mockResolvedValue([
       {
