+++ src/preload/index.ts
@@ -507,8 +507,6 @@ const api = {
       ipcRenderer.invoke(IPC.PLUGIN.CLEANUP_PROJECT_INJECTIONS, pluginId, projectPath),
     listOrphanedPluginIds: (projectPath: string, knownPluginIds: string[]): Promise<string[]> =>
       ipcRenderer.invoke(IPC.PLUGIN.LIST_ORPHANED_PLUGIN_IDS, projectPath, knownPluginIds),
-    registerManifest: (pluginId: string, manifest: any) =>
-      ipcRenderer.invoke(IPC.PLUGIN.REGISTER_MANIFEST, pluginId, manifest),
     refreshManifestFromDisk: (pluginId: string) =>
       ipcRenderer.invoke(IPC.PLUGIN.REFRESH_MANIFEST_FROM_DISK, pluginId),
   },
